Fix sql error about duplicated record during the login
Ldap user retrieval is using insensitive case comparison. So when the user gives a login
that have not the same case as indicated into the ldap, and when
this ldap login is mapped into the login field of the account, it may be different from
the login given by the user.

In this case, the login stored into the database account is the one given by ldap
and not by the user, so in the next login process, multiauth does not find the record
and tries to create a new record into the database. And it fails because of duplicated
keys.

Multiauth must check if the account does not already exists AFTER the provider
validate the authentication, to be sure to not create a record that already exists.

Refs #3

Some LDAP mapping may rewrite login